requirements.txt
Flask==1.1.2
key_generator==1.0.3
mysql_connector_repackaged==0.3.1
